test: split doc contract suites by ownership
