+++ test/extension-package-contract.test.mjs
@@ -22,35 +22,42 @@ test("package metadata exposes the extension entrypoint and root extension test 
   assert.deepEqual(packageJson.pi.skills, [".pi/skills"]);
 });
 
-test("extension README documents the command surface and runtime/build/test contract", async () => {
+test("extension README documents the supported command, install, and verification surfaces without exposing internal workflow seams", async () => {
   const readme = await readRepo("extension/README.md");
 
-  assert.match(readme, /defaults to help output/i);
-  assert.match(readme, /\/dev-loops status/i);
-  assert.match(readme, /pi-dev-loops status/i);
-  assert.match(readme, /concise readiness summary/i);
-  assert.match(readme, /\/dev-loops doctor/i);
-  assert.match(readme, /full diagnostic report/i);
-  assert.match(readme, /pi install git:github.com\/mfittko\/pi-dev-loops/i);
-  assert.match(readme, /pi install -l git:github.com\/mfittko\/pi-dev-loops/i);
-  assert.match(readme, /pi update git:github.com\/mfittko\/pi-dev-loops/i);
-  assert.match(readme, /Node[^\n]*>=20/i);
-  assert.match(readme, /source-loaded/i);
-  assert.match(readme, /package\.json` `pi\.skills`/i);
-  assert.match(readme, /\.pi\/agents\/\*\.agent\.md/i);
-  assert.match(readme, /~\/\.agents/i);
-  assert.match(readme, /single public workflow entry/i);
-  assert.match(readme, /readiness surface should not present them as separate user-facing checks/i);
-  assert.match(readme, /are removed; use `pi install` \/ `pi update` directly instead/i);
-  assert.doesNotMatch(readme, /\/skill:copilot-dev-loop|\/skill:copilot-autopilot/i);
-  assert.match(readme, /node --import tsx --test/i);
-  assert.match(readme, /does not yet claim a specific supported `gh` version/i);
-  assert.match(readme, /npm run verify/i);
-  assert.match(readme, /npm run test:extension/i);
-  assert.match(readme, /npm run test:dev-loop/i);
-  assert.match(readme, /npm run test:playwright:viewer/i);
-});
+  for (const commandPattern of [
+    /\/dev-loops status/i,
+    /\/dev-loops doctor/i,
+    /pi-dev-loops status/i,
+  ]) {
+    assert.match(readme, commandPattern);
+  }
 
+  for (const installPattern of [
+    /pi install git:github.com\/mfittko\/pi-dev-loops/i,
+    /pi install -l git:github.com\/mfittko\/pi-dev-loops/i,
+    /pi update git:github.com\/mfittko\/pi-dev-loops/i,
+  ]) {
+    assert.match(readme, installPattern);
+  }
+
+  for (const runtimePattern of [
+    /Node[^\n]*>=20/i,
+    /source-loaded/i,
+    /package\.json` `pi\.skills`/i,
+    /\.pi\/agents\/\*\.agent\.md/i,
+    /~\/\.agents/i,
+    /single public workflow entry/i,
+    /npm run verify/i,
+    /npm run test:extension/i,
+    /npm run test:dev-loop/i,
+    /npm run test:playwright:viewer/i,
+  ]) {
+    assert.match(readme, runtimePattern);
+  }
+
+  assert.doesNotMatch(readme, /\/skill:copilot-dev-loop|\/skill:copilot-autopilot/i);
+});
 
 test("required installed runtime contract docs are bundled once in the shared installed docs location", async () => {
   const extensionReadme = await readRepo("extension/README.md");
