+++ test/review-doc-contracts.test.mjs
@@ -1,0 +1,133 @@
+import {
+  assert,
+  fromRepoRoot,
+  parseFrontmatter,
+  readRepo,
+  readdir,
+  stat,
+  test,
+  USER_FACING_AGENT_SURFACE,
+} from "./imported-assets-helpers.mjs";
+
+test("copilot skill does not contain known imported blocker phrases", async () => {
+  const content = await readRepo("skills/copilot-dev-loop/SKILL.md");
+
+  assert.doesNotMatch(content, /repo-wiki/);
+  assert.doesNotMatch(content, /copilot-review-followup/);
+  assert.doesNotMatch(content, /async-review-fix-push/);
+});
+
+test("review agent does not hardcode reviewer identity or stale plan path", async () => {
+  const content = await readRepo("agents/review.agent.md");
+
+  assert.doesNotMatch(content, /mfittko/);
+  assert.doesNotMatch(content, /docs\/plans\//);
+});
+
+test("coordinator agent does not contain stale docs/plans path and requires fresh-context review briefings", async () => {
+  const content = await readRepo("agents/coordinator.agent.md");
+
+  assert.doesNotMatch(content, /docs\/plans\//);
+  assert.match(content, /Do not fork the parent session for review subagents/i);
+  assert.match(content, /concise briefing summary/i);
+  assert.match(content, /fresh context/i);
+});
+
+test("review workflow documents DRY/KISS/YAGNI as default pre-approval gate with explicit fallback requirement", async () => {
+  const [devLoopSkill, copilotSkill, reviewAgent, coordinatorAgent, reviewTemplate, reviewerGraph] = await Promise.all([
+    readRepo("skills/dev-loop/SKILL.md"),
+    readRepo("skills/copilot-dev-loop/SKILL.md"),
+    readRepo("agents/review.agent.md"),
+    readRepo("agents/coordinator.agent.md"),
+    readRepo("skills/dev-loop/templates/review.md"),
+    readRepo("docs/reviewer-loop-state-graph.md"),
+  ]);
+
+  const gateDocuments = [
+    ["skills/dev-loop/SKILL.md", devLoopSkill, /default pre-approval gate[\s\S]{0,200}\bDRY\b[\s\S]{0,80}\bKISS\b[\s\S]{0,80}\bYAGNI\b/i],
+    ["skills/copilot-dev-loop/SKILL.md", copilotSkill, /default pre-approval gate[\s\S]{0,200}\bDRY\b[\s\S]{0,80}\bKISS\b[\s\S]{0,80}\bYAGNI\b/i],
+    ["agents/review.agent.md", reviewAgent, /default pre-approval gate contract:[\s\S]{0,160}\bDRY\b[\s\S]{0,80}\bKISS\b[\s\S]{0,80}\bYAGNI\b/i],
+    ["agents/coordinator.agent.md", coordinatorAgent, /default pre-approval review fan-out must use the [\s\S]{0,40}\bDRY\b[\s\S]{0,40}\bKISS\b[\s\S]{0,40}\bYAGNI\b lenses/i],
+    ["skills/dev-loop/templates/review.md", reviewTemplate, /^## Default pre-approval gate \(DRY \/ KISS \/ YAGNI\)$/m],
+    ["docs/reviewer-loop-state-graph.md", reviewerGraph, /default pre-approval gate[\s\S]{0,200}\bDRY\b[\s\S]{0,80}\bKISS\b[\s\S]{0,80}\bYAGNI\b/i],
+  ];
+
+  for (const [label, content, gatePhraseWithLenses] of gateDocuments) {
+    assert.match(content, gatePhraseWithLenses, `${label} should keep the gate phrasing and lens names aligned`);
+  }
+
+  for (const [label, content] of [
+    ["skills/dev-loop/SKILL.md", devLoopSkill],
+    ["skills/copilot-dev-loop/SKILL.md", copilotSkill],
+    ["agents/review.agent.md", reviewAgent],
+    ["agents/coordinator.agent.md", coordinatorAgent],
+    ["docs/reviewer-loop-state-graph.md", reviewerGraph],
+  ]) {
+    assert.match(
+      content,
+      /review-complete, approval-ready, merge-ready, or ready for final handoff/i,
+      `${label} should keep the gate boundary wording aligned`,
+    );
+  }
+
+  assert.match(reviewTemplate, /fallback note:[^\n]*if parallel execution of the three review lenses is impractical/i);
+  assert.match(devLoopSkill, /if parallel execution is impractical[\s\S]*run all three lenses sequentially and explicitly record why parallel execution was impractical/i);
+  assert.match(copilotSkill, /fresh context and in parallel when practical/i);
+  assert.match(copilotSkill, /if parallel execution is impractical[\s\S]*still run all three lenses and explicitly record the limitation/i);
+  assert.match(reviewAgent, /if parallel execution is impractical[\s\S]*still cover all three lenses and explicitly record the limitation/i);
+  assert.match(coordinatorAgent, /default to three focused lenses \(DRY, KISS, YAGNI\) and run them in parallel when practical/i);
+  assert.match(coordinatorAgent, /if parallel execution is impractical[\s\S]*still run all three lenses and record that limitation explicitly/i);
+  assert.match(reviewerGraph, /workflow lenses that reviewer\s+runs must cover for the change/i);
+  assert.match(reviewerGraph, /do not replace the state machine's supported\s+review-angle taxonomy/i);
+});
+
+test("reviewer-loop contract documents submitted-review handoff and explicit external waits", async () => {
+  const [reviewerGraph, scriptsReadme] = await Promise.all([
+    readRepo("docs/reviewer-loop-state-graph.md"),
+    readRepo("scripts/README.md"),
+  ]);
+
+  assert.match(reviewerGraph, /A pure internal reviewer pass must end in a concrete review result boundary \(`submitted_review`\)/i);
+  assert.match(reviewerGraph, /If a wait state is used, it must be an explicit named external-participant boundary/i);
+  assert.match(reviewerGraph, /A new review request after fixes starts a new reviewer-pass context \(`review_requested`\)/i);
+  assert.match(scriptsReadme, /reviewer `submitted_review`\s+as outer-loop-owned `continue_wait` states at explicit external\/handoff boundaries/i);
+  assert.match(scriptsReadme, /preserves compatibility for reviewer `waiting_for_author_followup` and `waiting_for_re_request`\s+as legacy named external-wait boundaries/i);
+});
+
+test("dev-loop skill documents opt-in Playwright smoke harnesses for UI slices", async () => {
+  const devLoopSkill = await readRepo("skills/dev-loop/SKILL.md");
+
+  assert.match(devLoopSkill, /user-facing HTML\/UI\/component slices when the user opts in/i);
+  assert.match(devLoopSkill, /Playwright WebKit plus screenshot capture/i);
+  assert.match(devLoopSkill, /wire it into CI once it becomes required validation for that slice/i);
+});
+
+
+test("CI gates the Playwright WebKit smoke behind inspect-run viewer change detection and uses Node24-ready first-party actions", async () => {
+  const [ciWorkflow, readme] = await Promise.all([
+    readRepo(".github/workflows/ci.yml"),
+    readRepo("README.md"),
+  ]);
+
+  assert.match(ciWorkflow, /^\s{2}changes:\s*$/m);
+  assert.match(ciWorkflow, /^\s{2}verify:\s*$/m);
+  assert.match(ciWorkflow, /^\s{2}viewer-smoke:\s*$/m);
+  assert.match(ciWorkflow, /fetch-depth:\s*0/i);
+  assert.match(ciWorkflow, /actions\/checkout@v5/i);
+  assert.match(ciWorkflow, /actions\/setup-node@v5/i);
+  assert.match(ciWorkflow, /actions\/cache@v5/i);
+  assert.match(ciWorkflow, /changes:[\s\S]*Set up Node\.js[\s\S]*node-version:\s*24/i);
+  assert.match(ciWorkflow, /GITHUB_OUTPUT="\$GITHUB_OUTPUT" node scripts\/loop\/inspect-run-viewer-ci-changes\.mjs \.inspect-run-viewer-changed-files\.txt/i);
+  assert.doesNotMatch(ciWorkflow, /inspect_run_viewer_relevant_paths_json/i);
+  assert.match(ciWorkflow, /viewer-smoke:[\s\S]*needs:[\s\S]*- changes/i);
+  assert.match(ciWorkflow, /viewer-smoke:[\s\S]*if:\s*needs\.changes\.outputs\.inspect_run_viewer\s*==\s*'true'/i);
+  assert.match(ciWorkflow, /viewer-smoke:[\s\S]*path:\s*\$\{\{\s*env\.PLAYWRIGHT_BROWSERS_PATH\s*\}\}/i);
+  assert.match(ciWorkflow, /PLAYWRIGHT_BROWSERS_PATH:\s*\$\{\{\s*github\.workspace\s*\}\}\/\.cache\/ms-playwright/i);
+  assert.match(ciWorkflow, /key:\s*\$\{\{\s*runner\.os\s*\}\}-playwright-webkit-\$\{\{\s*hashFiles\('package-lock\.json'\)\s*\}\}/i);
+  assert.match(ciWorkflow, /viewer-smoke:[\s\S]*npm run test:playwright:viewer/i);
+  assert.match(ciWorkflow, /verify:[\s\S]*npm run verify/i);
+
+  assert.match(readme, /workspace-local Playwright WebKit/i);
+  assert.match(readme, /small changed-files gate plus parallel `verify` and conditional `viewer-smoke` jobs/i);
+  assert.match(readme, /run only when files in the bounded inspect-run viewer surface change/i);
+});
